Refactor modules and fix discovry bug
