+++ src/utils.rs
@@ -1,0 +1,21 @@
+use multiaddr::Multiaddr;
+use std::net::SocketAddr;
+
+/// Change multiaddr to socketaddr
+pub fn multiaddr_to_socketaddr(addr: &Multiaddr) -> Result<SocketAddr, ()> {
+    use multiaddr::Protocol;
+
+    let mut iter = addr.iter();
+    let proto1 = iter.next().ok_or(())?;
+    let proto2 = iter.next().ok_or(())?;
+
+    if iter.next().is_some() {
+        return Err(());
+    }
+
+    match (proto1, proto2) {
+        (Protocol::Ip4(ip), Protocol::Tcp(port)) => Ok(SocketAddr::new(ip.into(), port)),
+        (Protocol::Ip6(ip), Protocol::Tcp(port)) => Ok(SocketAddr::new(ip.into(), port)),
+        _ => Err(()),
+    }
+}
