+++ tests/test_dial.rs
@@ -1,11 +1,11 @@
 use futures::prelude::Stream;
 use p2p::{
     builder::ServiceBuilder,
+    context::{ServiceContext, SessionContext},
     multiaddr::Multiaddr,
-    service::{ProtocolMeta, ServiceContext, ServiceProtocol, SessionContext},
-    service::{Service, ServiceEvent, ServiceHandle},
-    session::ProtocolId,
-    SecioKeyPair, SessionType,
+    service::{Service, ServiceEvent},
+    traits::{ProtocolMeta, ServiceHandle, ServiceProtocol},
+    ProtocolId, SecioKeyPair, SessionType,
 };
 use std::{thread, time::Duration};
 use tokio::codec::LengthDelimitedCodec;
